ir: translation_unit::is_empty should work without environment.
I noticed that some code failed trying to call
translation_unit::is_empty in the absence of the instance of the
abigail::ir::environment type used to create the IR.

This is because translation_unit::is_empty was trying to create
something.  That's odd.  It shouldn't need to create anything to test
for its emptiness.  Fixed thus.

	* src/abg-ir.cc (translation_unit::is_empty): If there is no
	global scope, then we know its empty.  No need to create one.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>