Fix IR comparison result caching and canonical type propagation tracking
Caching the result of IR comparison cannot happen on IR nodes that are
being compared in the context of type canonicalization if one of the
nodes is the target of canonical type propagation.  This is especially
true if the recursive IR node which comparison "temporarily" did yield
true (to avoid an infinite loop) at least until the comparison of the
rest of the sub-tree of the recursive type is done.  In that case, we
should not cache the result the comparison as it might change later.

The patch adds a way to track recursive types so that we know when not
to cache their comparison result.

As we now have a facility to track recursive types during canonical
type propagation (rather than just types that depend on recursive
types) we can be a bit more precise when confirming or cancelling
types that have been subject to canonical type propagation.

Also the patch cleans up the detection of comparison cycle to make it
more typesafe so that the macro
RETURN_TRUE_IF_COMPARISON_CYCLE_DETECTED can be used for class_decl
and union_decl, not just class_or_union.  It makes the code more
readable/maintainable is the equals overload for class_decl and
union_decl all have their proper call to
RETURN_TRUE_IF_COMPARISON_CYCLE_DETECTED.  The same is true for
mark_types_as_being_compared.

Doing that cleans up the detection of recursive types as well as the
types that depends on those recursive types.

	* src/abg-ir-priv.h (environment::priv::recursive_types_): Define
	new data member.
	(environment::priv::cache_type_comparison_result): Cache results
	only non-recursive and not dependant types, or when the result is
	"false".  In all these cases, the result, once cached, will not
	change.
	(environment::priv::mark_dependant_types_compared_until): Mark
	types as recursive at the same time others are marked as
	dependant.
	(environment::priv::{is_recursive_type, set_is_not_recursive}):
	Define new member functions.
	(environment::priv::{confirm_ct_propagation,
	cancel_ct_propagation}): Confirming canonical type propagation
	should happen for recursive types as well as their dependant
	types.
	* src/abg-ir.cc (return_comparison_result): Keep up with the
	book-keeping at all time when type canonicalization process is
	on-going.  Whenever we expect types that depends on recursive
	types, expect recursive types too, obviously.
	(type_base::get_canonical_type_for): Do not erase the comparison
	result cache between the canonicalization of two different types.
	(is_comparison_cycle_detected)
	(mark_types_as_being_compared, unmark_types_as_being_compared):
	Define new overloads for class_decl;
	(equals): In the overloads for class_decl and union_decl, use
	RETURN_COMPARISON_RESULT and mark_types_as_being_compared without
	casting it to class_or_union.
	* tests/data/test-abidiff/test-PR18791-report0.txt: Adjust.
	* tests/data/test-diff-dwarf/PR25058-liblttng-ctl-report-1.txt:
	Adjust.
	* tests/data/test-diff-pkg/nss-3.23.0-1.0.fc23.x86_64-report-0.txt:
	Adjust.
	* tests/data/test-diff-pkg/spice-server-0.12.4-19.el7.x86_64-0.12.8-1.el7.x86_64-report-2.txt:
	Adjust.
	* tests/data/test-diff-pkg/tbb-4.1-9.20130314.fc22.x86_64--tbb-4.3-3.20141204.fc23.x86_64-report-0.txt:
	Adjust.
	* tests/data/test-diff-pkg/tbb-4.1-9.20130314.fc22.x86_64--tbb-4.3-3.20141204.fc23.x86_64-report-1.txt:
	Adjust.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>