ir: Support cloning data members of unions
When cloning a data member, var_decl::clone wrongly assumes we are
looking at a data member of a class.  It could also be a data member
of an union.  Fixed thus.

This fixes crashes when self comparing Fedora 36 packages of the qt6
group with the command:

    $ fedabipkgdiff --self-compare --from fc36 qt6-qtsensors

There are lot of other packages from the qt group that were making
this command fail, but I am not mentioning them here for the sake
conciseness.

	* src/abg-ir.cc (var_decl::clone): Support cloning data members
	for unions.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>