Bug 29650 - Caching class comparison result potentially too early
When structurally comparing two classes T and T' the overload of the
equals() function for abigail::ir::class_decl calls the overload of
equals() for abigail::ir::class_or_union to compare the data members
of the class.  If data members are equal, that later call caches the
result of comparing the data-members-only sub-object of T and T'.
That caching appears as if it's the result of comparing all of T and
T' that was cached, leading to misleading results down the road.

Result caching should not take place until the end of fulling
comparing T and T'.  Fixed thus.

	* src/abg-ir.cc (equal): In the overload of class_or_union do not
	cache the result comparing just the data members sub-types of of
	classes. In the overload for class_decl, put cycle detection
	management code /after/ the call to equals for class_or_union,
	because that called function does perform the cycle detection
	management as well; otherwise, that introduces an unwarranted
	redundancy.  In the overload of equals for union_decl, cache the
	result of the comparison.
	* tests/data/test-abidiff/test-PR18791-report0.txt: Adjust.
	* tests/data/test-diff-pkg/tbb-4.1-9.20130314.fc22.x86_64--tbb-4.3-3.20141204.fc23.x86_64-report-0.txt:
	Adjust.
	* tests/data/test-diff-pkg/tbb-4.1-9.20130314.fc22.x86_64--tbb-4.3-3.20141204.fc23.x86_64-report-1.txt:
	Adjust.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>