abg-ir: add missing else
In 'bind_function_type_life_time' function of 'src/abg-ir.cc', the code
obtains the member 'const environment*' of 'class function_type', that
is, the 'e' variable in below code. And assure the obtained
'environment*' is same as the 'const environment*' of the
'class translation_unit', that is, the'env' variable in below code:

/* src/abg-ir.cc begin */
1  void
2  translation_unit::bind_function_type_life_time(function_type_sptr ftype)
3  {
4    ...
5    const environment* env = get_environment();
6    ...
7    if (const environment* e = ftype->get_environment())
8      ABG_ASSERT(env == e);
9    ftype->set_environment(const_cast<environment*>(env));
10
11   if (const translation_unit* existing_tu = ftype->get_translation_unit())
12     ABG_ASSERT(existing_tu == this);
13   else
14     ftype->set_translation_unit(const_cast<translation_unit*>(this));
15    ...
/* src/abg-ir.cc end */

There was a missing 'else' between the 'line 8' and line 9', as the
explicit 'else' at the 'line 13'. Without the 'else' between the
'line 8' and line 9', there will be a redundant assignment at 'line 9'
under the condition when the 'env' is equal to 'e'.
This patch add the missing 'else' between the 'line 8' and 'line 9'.

        * src/abg-ir.cc (bind_function_type_life_time): add missing
        else

Signed-off-by: Xiaole He <hexiaole@kylinos.cn>
Tested-by: Xiaole He <hexiaole@kylinos.cn>