rhbz2114909 - Refer to changed base classes using their non-qualified names
This was filled at https://bugzilla.redhat.com/show_bug.cgi?id=2114909

When analysing the edit script representing the changes in the base
base classes of a given class, libabigail refers to the changed base
classes using their names.

The problem is that there can be several changed base classes that
have the same name, even though their /qualified/ names are
different.  That was wreaking having havoc down the road.

To solve the problem, use the qualified name of the base classes,
instead of their simple name.

Note that this change did also ameliorate other existing change
reports in the test suite.

While testing the fix, I realized that class_decl::base_spec doesn't
set its qualified name.  That leads to the fact that it's the
non-qualified name that is used for sorting the base class names in
the change reports.  This can lead to instabilities, depending on the
platform, in test output.  This is also fixed.

Fixed thus.

	* src/abg-comparison.cc
	(class_diff::ensure_lookup_tables_populated): Refer to the added
	or removed base class using its qualified name.
	* src/abg-ir.cc (class_decl::base_spec::base_spec): Set the
	qualified name of the base_spec.
	* tests/data/test-abidiff-exit/test-rhbz2114909-report-1.txt: New
	reference test output.
	* tests/data/test-abidiff-exit/test-rhbz2114909-v{0,1}.o: New
	input test binaries.
	* tests/data/test-abidiff-exit/test-rhbz2114909-v{0,1}.cc: Source
	code of the binaries, above.
	* tests/data/Makefile.am: Add the new test materials above to
	source distribution.
	* tests/test-abidiff-exit.cc (in_out_specs): Add the new tests to
	this harness.
	* tests/data/test-diff-pkg/libsigc++-2.0-0c2a_2.4.0-1_amd64--libsigc++-2.0-0v5_2.4.1-1ubuntu2_amd64-report-0.txt: Adjust.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>